+++ unicode_fol_kit/__init__.py
@@ -26,6 +26,7 @@ from .fol import (
 )
 from .atp import (
     formulas_are_equivalent, check_logical_entailment,
+    check_logical_entailment_vampire,
     is_satisfiable, is_valid, get_model,
     fuzzy_is_satisfiable, fuzzy_is_valid, fuzzy_get_model,
     to_z3_arith, is_satisfiable_arith, is_valid_arith, get_model_arith,
@@ -52,6 +53,7 @@ __all__ = [
     "Z3Env",
     "NamingError", "ParsingError",
     "formulas_are_equivalent", "check_logical_entailment",
+    "check_logical_entailment_vampire",
     "SortedQuantifier", "SortedConstant",
     "WeakConjunction", "WeakDisjunction",
     "StrongConjunction", "StrongDisjunction",