+++ unicode_fol_kit/atp/__init__.py
@@ -1,5 +1,6 @@
 from .z3_equivalence import formulas_are_equivalent
 from .prover9_entailment import check_logical_entailment
+from .vampire_entailment import check_logical_entailment_vampire
 from .z3_models import is_satisfiable, is_valid, get_model
 from .z3_fuzzy import (
     fuzzy_is_satisfiable, fuzzy_is_valid, fuzzy_get_model, degree_expr,
@@ -12,6 +13,7 @@ from .resolution import to_clauses, refute, prove, is_valid_resolution
 __all__ = [
     "formulas_are_equivalent",
     "check_logical_entailment",
+    "check_logical_entailment_vampire",
     "is_satisfiable", "is_valid", "get_model",
     "fuzzy_is_satisfiable", "fuzzy_is_valid", "fuzzy_get_model", "degree_expr",
     "to_z3_arith", "is_satisfiable_arith", "is_valid_arith", "get_model_arith", "ArithEnv",